Moved pricing table into a PricingTab folder
