+++ src/components/CostAndPricing/CostAndPricing.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import { PricingTable } from './PricingTable';
-import { quote as sampleQuote } from './sample-data';
+import { PricingTable } from './PricingTab/PricingTable';
+import { quote as sampleQuote } from './PricingTab/sample-data';
 
 export default function CostAndPricing() {
   const [quote, setQuote] = useState(sampleQuote);
